made new class for link popup
