+++ MealInfoDisplay.py
@@ -2,11 +2,9 @@ from tkinter import *
 from tkinter import messagebox
 from tkinter import ttk
 from PlanPage import PlanPage
-import webbrowser
-from PIL import Image
-from PIL import ImageTk
 from MealFileSaver import MealFileSaver
 from idlelib.tooltip import Hovertip
+from MealDetailPopup import MealDetailPopup
 
 class MealInfoDisplay(PlanPage):
     def __init__(self, outer, mealcreator):
@@ -74,50 +72,7 @@ class MealInfoDisplay(PlanPage):
         reroll_tip = Hovertip(self.reroll_button, "Re-Roll")
 
     def show_meal_details(self, mealday, meal):
-        # open a message box that shows the ingredients and the recipe
-
-        meal_choice = mealday.get_choice(meal)
-
-        information = meal_choice.format_meal_ingredients()
-        
-        if meal_choice.link != None:
-            information += "\n\nRecipe Link:\n"
-            self.details_popup= Toplevel(bg="white")
-            self.details_popup.title("Details for " + meal_choice.name.title())
-            
-            bottom_frame = Frame(self.details_popup, bg="gray95")
-            bottom_frame.pack(side=BOTTOM, fill=X)
-            ttk.Button(bottom_frame, text="OK", width=12, default="active", command=self.details_popup.destroy).pack(anchor='se', pady = 10, padx = 15)
-            
-            left_frame = Frame(self.details_popup, width = 40, bg="white")
-            left_frame.pack(side=LEFT, fill=Y, padx=(15, 15))
-            self.info_icon = Image.open("./information_icon.png")
-            self.info_icon = (self.info_icon).resize((50,50))
-            self.info_icon = ImageTk.PhotoImage(self.info_icon)
-            self.info_icon_label = Label(left_frame, image=self.info_icon, bg="white", padx=10, pady=20)
-            self.info_icon_label.pack(anchor='n')
-
-            Label(self.details_popup, text= information, justify = LEFT, bg="white", wraplength=300).pack(anchor = "w", padx=(0, 20), pady=(15,0))
-            self.create_hyperlink(meal_choice.link)
-            return
-        
-        if meal_choice.recipe != None:
-            information += "\n\nRecipe: \n\n"
-            information += meal_choice.recipe
-            information += "\n"
-
-        messagebox.showinfo("Details for " + meal_choice.name.title(), information)
-
-    def open_hyperlink(self, url):
-        webbrowser.open_new_tab(url)
-
-    #Create a Label to display the link
-    def create_hyperlink(self, link_text):
-        link = Label(self.details_popup, text=link_text, fg="blue", cursor="hand2", bg="white", wraplength=300, justify=LEFT)
-        link.bind("<Button-1>", lambda e:
-        self.open_hyperlink(link_text))
-        link.pack(pady = (0, 15), padx=(0,20))
-
+        MealDetailPopup(mealday, meal)
 
     def reroll_meal(self, mealday, meal, label):
         self.meal_creator.select_meal(mealday, meal)
